feat(insight): KpiService impl + 12 unit tests + 3 Feign clients

+++ springboot3/ruoyi-modules/opc-billing/src/main/java/com/ruoyi/opc/billing/controller/OpcBillingController.java
@@ -135,7 +135,7 @@ public class OpcBillingController extends BaseController {
      */
     @ExceptionHandler(OpcException.class)
     public R<Void> handleOpcException(OpcException e) {
-        return R.fail(e.getCode(), e.getMessage());
+        return R.fail(e.getCode().intValue(), e.getMessage());
     }
 
     @lombok.Data
