+++ springboot3/ruoyi-modules/opc-finance/src/main/java/com/ruoyi/opc/finance/controller/OpcFinanceController.java
@@ -175,7 +175,7 @@ public class OpcFinanceController extends BaseController {
      */
     @ExceptionHandler(OpcException.class)
     public R<Void> handleOpcException(OpcException e) {
-        return R.fail(e.getCode(), e.getMessage());
+        return R.fail(e.getCode().intValue(), e.getMessage());
     }
 
 }
